Al terminal logging rewrite

+++ hyperscale/core/engines/client/time_parser.py
@@ -11,7 +11,7 @@ class TimeParser:
             "d": "days",
             "w": "weeks",
         }
-        self.time = int(
+        self.time = float(
             timedelta(
                 **{
                     self.UNITS.get(m.group("unit").lower(), "seconds"): float(
